feat: implement integrations among artist, album and track

+++ src/artist/artist.service.ts
@@ -1,10 +1,16 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { ArtistRepository } from './artist.repository';
 import { Artist } from '../models/artist.entity';
+import { AlbumService } from 'src/album/album.service';
+import { TrackService } from 'src/track/track.service';
 
 @Injectable()
 export class ArtistService {
-  constructor(private readonly artistRepository: ArtistRepository) {}
+  constructor(
+    private readonly artistRepository: ArtistRepository,
+    private readonly trackService: TrackService,
+    private readonly albumService: AlbumService,
+  ) {}
 
   async createArtist(artist: Omit<Artist, 'id'>): Promise<Artist> {
     return this.artistRepository.create(artist);
@@ -26,6 +32,14 @@ export class ArtistService {
   }
 
   async deleteArtist(id: string): Promise<void> {
-    return this.artistRepository.delete(id);
+    const artist = await this.artistRepository.findById(id);
+    if (!artist) {
+      throw new NotFoundException(`Artist with ID ${id} not found`);
+    }
+
+    await this.artistRepository.delete(id);
+
+    await this.trackService.setArtistIdToNullForTrack(id);
+    await this.albumService.setArtistIdToNullForAlbum(id);
   }
 }
